+++ takumi-image-response/src/backends/wasm.ts
@@ -1,7 +1,8 @@
+import { fetchResources } from "@takumi-rs/helpers";
 import { type FromJsxOptions, fromJsx } from "@takumi-rs/helpers/jsx";
 import init, {
   type ByteBuf,
-  collectNodeFetchTasks,
+  extractResourceUrls,
   type Font,
   type InitInput,
   Renderer,
@@ -130,20 +131,10 @@ function createStream(component: ReactNode, options: ImageResponseOptions) {
         const node = await fromJsx(component, options.jsx);
 
         if (!options.fetchedResources) {
-          const urls = collectNodeFetchTasks(node);
+          const urls = extractResourceUrls(node);
 
           if (urls.length > 0) {
-            options.fetchedResources = new Map(
-              await Promise.all(
-                urls.map(
-                  async (url) =>
-                    [
-                      url,
-                      await fetch(url).then((r) => r.arrayBuffer()),
-                    ] as const,
-                ),
-              ),
-            );
+            options.fetchedResources = await fetchResources(urls);
           }
         }