+++ takumi-napi-core/src/renderer.rs
@@ -1,4 +1,3 @@
-use lru::LruCache;
 use napi::bindgen_prelude::*;
 use napi_derive::napi;
 use takumi::{
@@ -6,28 +5,18 @@ use takumi::{
   layout::node::NodeKind,
   parley::{FontWeight, GenericFamily, fontique::FontInfoOverride},
   rendering::ImageOutputFormat,
-  resources::{
-    image::{ImageSource, load_image_source_from_bytes},
-    task::FetchTask,
-  },
+  resources::image::load_image_source_from_bytes,
 };
 
 use crate::{
-  FetchFn, FontInput, buffer_from_object, buffer_slice_from_object, deserialize_with_tracing,
+  FontInput, buffer_from_object, buffer_slice_from_object, deserialize_with_tracing,
   load_font_task::LoadFontTask, map_error, put_persistent_image_task::PutPersistentImageTask,
   render_animation_task::RenderAnimationTask, render_task::RenderTask,
 };
-use std::{
-  num::NonZeroUsize,
-  sync::{Arc, Mutex},
-};
-
-pub(crate) type ResourceCache = Option<Arc<Mutex<LruCache<FetchTask, Arc<ImageSource>>>>>;
 
 #[napi]
 pub struct Renderer {
   global: GlobalContext,
-  resources_cache: ResourceCache,
 }
 
 #[napi(object)]
@@ -43,10 +32,8 @@ pub struct RenderOptions<'env> {
   pub quality: Option<u8>,
   /// Whether to draw debug borders.
   pub draw_debug_border: Option<bool>,
-  /// The fetch function to use to fetch resources.
-  /// @default global fetch function
-  #[napi(ts_type = "typeof fetch")]
-  pub fetch: Option<FetchFn<'env>>,
+  /// The fetched resources to use.
+  pub fetched_resources: Option<Vec<ImageSource<'env>>>,
   /// The device pixel ratio.
   /// @default 1.0
   pub device_pixel_ratio: Option<f64>,
@@ -98,7 +85,7 @@ impl From<OutputFormat> for ImageOutputFormat {
 }
 
 #[napi(object)]
-pub struct PersistentImage<'ctx> {
+pub struct ImageSource<'ctx> {
   pub src: String,
   #[napi(ts_type = "Uint8Array | ArrayBuffer")]
   pub data: Object<'ctx>,
@@ -107,11 +94,14 @@ pub struct PersistentImage<'ctx> {
 #[napi(object)]
 #[derive(Default)]
 pub struct ConstructRendererOptions<'ctx> {
-  pub persistent_images: Option<Vec<PersistentImage<'ctx>>>,
+  /// The images that needs to be preloaded into the renderer.
+  pub persistent_images: Option<Vec<ImageSource<'ctx>>>,
+  /// The fonts being used.
   #[napi(ts_type = "Font[] | undefined")]
   pub fonts: Option<Vec<Object<'ctx>>>,
+  /// Whether to load the default fonts.
+  /// If `fonts` are provided, this will be `false` by default.
   pub load_default_fonts: Option<bool>,
-  pub resource_cache_capacity: Option<u32>,
 }
 
 const EMBEDDED_FONTS: &[(&[u8], &str, GenericFamily)] = &[
@@ -127,8 +117,6 @@ const EMBEDDED_FONTS: &[(&[u8], &str, GenericFamily)] = &[
   ),
 ];
 
-const DEFAULT_RESOURCE_CACHE_CAPACITY: u32 = 8;
-
 #[napi]
 impl Renderer {
   #[napi(constructor)]
@@ -138,10 +126,6 @@ impl Renderer {
     let load_default_fonts = options
       .load_default_fonts
       .unwrap_or_else(|| options.fonts.is_none());
-
-    let resource_cache_capacity = options
-      .resource_cache_capacity
-      .unwrap_or(DEFAULT_RESOURCE_CACHE_CAPACITY);
 
     let mut global = GlobalContext::default();
 
@@ -192,18 +176,7 @@ impl Renderer {
       }
     }
 
-    let mut renderer = Self {
-      global,
-      resources_cache: if resource_cache_capacity > 0 {
-        Some(Arc::new(Mutex::new(LruCache::new(
-          NonZeroUsize::new(resource_cache_capacity as usize).ok_or(Error::from_reason(
-            "Resource cache capacity must be greater than 0",
-          ))?,
-        ))))
-      } else {
-        None
-      },
-    };
+    let renderer = Self { global };
 
     if let Some(images) = options.persistent_images {
       for image in images {
@@ -220,16 +193,9 @@ impl Renderer {
     Ok(renderer)
   }
 
+  /// @deprecated This function does nothing.
   #[napi]
-  pub fn purge_resources_cache(&self) -> Result<()> {
-    if let Some(resource_cache) = self.resources_cache.as_ref() {
-      let mut lock = resource_cache.lock().map_err(map_error)?;
-
-      lock.clear();
-    }
-
-    Ok(())
-  }
+  pub fn purge_resources_cache(&self) {}
 
   #[napi(
     ts_args_type = "src: string, data: Uint8Array | ArrayBuffer, signal?: AbortSignal",
@@ -321,13 +287,7 @@ impl Renderer {
     let node: NodeKind = deserialize_with_tracing(source)?;
 
     Ok(AsyncTask::with_optional_signal(
-      RenderTask::from_options(
-        env,
-        node,
-        options.unwrap_or_default(),
-        &self.resources_cache,
-        &self.global,
-      )?,
+      RenderTask::from_options(env, node, options.unwrap_or_default(), &self.global)?,
       signal,
     ))
   }