+++ takumi/src/rendering/inline_drawing.rs
@@ -34,13 +34,32 @@ fn draw_glyph_run<I: GenericImageView<Pixel = Rgba<u8>>>(
   let run = glyph_run.run();
   let metrics = run.metrics();
 
+  if glyph_run.style().brush.background_color.0[3] > 0 {
+    let bg_height = metrics.ascent + metrics.descent;
+    let bg_y = glyph_run.baseline() - metrics.ascent;
+
+    canvas.fill_color(
+      Size {
+        width: glyph_run.advance(),
+        height: bg_height,
+      },
+      glyph_run.style().brush.background_color,
+      BorderProperties::default(),
+      context.transform
+        * Affine::translation(
+          layout.border.left + layout.padding.left + glyph_run.offset(),
+          layout.border.top + layout.padding.top + bg_y,
+        ),
+    );
+  }
+
   // decoration underline should not overlap with the glyph descent part,
   // as a temporary workaround, we draw the decoration under the glyph.
   if decoration_line.contains(&TextDecorationLine::Underline) {
     draw_decoration(
       canvas,
       glyph_run,
-      style.text_decoration_color,
+      glyph_run.style().brush.decoration_color,
       glyph_run.baseline() - metrics.underline_offset,
       glyph_run.run().font_size() / 18.0,
       layout,
@@ -71,8 +90,7 @@ fn draw_glyph_run<I: GenericImageView<Pixel = Rgba<u8>>>(
         layout,
         image_fill,
         context.transform,
-        context.opacity,
-        glyph_run.style(),
+        glyph_run.style().brush.color,
         palette,
       )?;
     }
@@ -85,7 +103,7 @@ fn draw_glyph_run<I: GenericImageView<Pixel = Rgba<u8>>>(
     draw_decoration(
       canvas,
       glyph_run,
-      style.text_decoration_color,
+      glyph_run.style().brush.decoration_color,
       offset,
       size,
       layout,
@@ -97,7 +115,7 @@ fn draw_glyph_run<I: GenericImageView<Pixel = Rgba<u8>>>(
     draw_decoration(
       canvas,
       glyph_run,
-      style.text_decoration_color,
+      glyph_run.style().brush.decoration_color,
       glyph_run.baseline() - metrics.ascent - metrics.underline_offset,
       glyph_run.run().font_size() / 18.0,
       layout,
@@ -114,7 +132,7 @@ pub(crate) fn draw_inline_box<N: Node<N>>(
   canvas: &mut Canvas,
   transform: Affine,
 ) -> Result<()> {
-  if node.context.opacity == 0 {
+  if node.context.style.opacity.0 == 0.0 {
     return Ok(());
   }